requirements.txt
alembic==1.5.8
aniso8601==9.0.1
appdirs==1.4.4
attrs==20.3.0
blinker==1.4
botocore==1.20.102
certifi==2020.12.5
chardet==3.0.4
Click==7.0
coverage==5.4
Flask==1.1.1
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
jmespath==0.10.0
greenlet==1.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
mongoengine==0.23.0
numpy==1.21.0
pycodestyle==2.7.0
regex==2021.4.4
s3transfer==0.3.7
snowballstemmer==2.1.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
psycopg2-binary==2.8.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
requests==2.22.0
six==1.13.0
urllib3==1.25.8
Werkzeug==1.0.0
boto3==1.17.49
dnspython==1.16.0
pymongo==3.11.4
pandas==1.2.4
XlsxWriter==1.4.3
marshmallow==3.11.1
uWSGI==2.0.20
openai==0.27.8
chromadb==0.4.0
tiktoken==0.4.0
unstructured==0.8.1